dev/experimental-utils/imgcat/pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "imgcat"
version = "0.1.0"
description = "An image viewer for terminal using Kitty graphics protocol"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "Pillow>=10.0.0",
    "click>=8.1.0",
    "cairosvg>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
imgcat = "imgcat.cli:main"

[tool.setuptools]
packages = ["imgcat"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
